refactor: separate webview resources and optimize GraphService
This commit significantly improves the maintainability, security, and performance of the graph view feature:

**Separate Webview Resources:**
- Extracted HTML, CSS, and JavaScript from string literal in graphView.ts into separate files (src/webview/graph.html, graph.css, graph.js)
- Updated graphView.ts to load resources using panel.webview.asWebviewUri() for secure resource loading
- Improved code maintainability with proper syntax highlighting, linting, and auto-formatting support

**Bundle vis-network Locally:**
- Added vis-network as a project dependency instead of loading from CDN
- Eliminates offline failure and security risks associated with external CDN
- Updated .vscodeignore to include vis-network/dist/** and src/webview/** in extension package

**Optimize GraphService with Caching:**
- Added private graphData cache to GraphService to avoid redundant graph rebuilds
- Updated getGraphStats() and searchNodes() to use cached data when available
- Added invalidateCache() method for cache management
- Eliminates duplicate buildGraph() calls during graph initialization and refresh

These changes result in:
- Better code organization and maintainability
- Improved security by bundling dependencies locally
- Enhanced performance by eliminating redundant graph builds
- Offline functionality for graph visualization
