+++ .vscodeignore
@@ -2,7 +2,9 @@
 .gitignore
 tsconfig.json
 src/**
+!src/webview/**
 node_modules/**
+!node_modules/vis-network/dist/**
 *.vsix
 pnpm-lock.yaml
 .DS_Store
