+++ src/webview/graph.css
@@ -1,0 +1,189 @@
+body {
+    margin: 0;
+    padding: 0;
+    font-family: var(--vscode-font-family);
+    background-color: var(--vscode-editor-background);
+    color: var(--vscode-editor-foreground);
+    overflow: hidden;
+}
+
+#container {
+    display: flex;
+    flex-direction: column;
+    height: 100vh;
+}
+
+#toolbar {
+    background-color: var(--vscode-sideBar-background);
+    border-bottom: 1px solid var(--vscode-panel-border);
+    padding: 12px;
+    display: flex;
+    gap: 12px;
+    align-items: center;
+    flex-wrap: wrap;
+}
+
+#stats {
+    background-color: var(--vscode-badge-background);
+    color: var(--vscode-badge-foreground);
+    padding: 4px 8px;
+    border-radius: 3px;
+    font-size: 12px;
+    display: flex;
+    gap: 16px;
+}
+
+.stat-item {
+    display: flex;
+    gap: 4px;
+}
+
+.stat-label {
+    opacity: 0.8;
+}
+
+#controls {
+    display: flex;
+    gap: 8px;
+    flex-wrap: wrap;
+    align-items: center;
+}
+
+input, select, button {
+    background-color: var(--vscode-input-background);
+    color: var(--vscode-input-foreground);
+    border: 1px solid var(--vscode-input-border);
+    padding: 6px 12px;
+    border-radius: 2px;
+    font-size: 13px;
+    font-family: var(--vscode-font-family);
+}
+
+input:focus, select:focus {
+    outline: 1px solid var(--vscode-focusBorder);
+}
+
+button {
+    background-color: var(--vscode-button-background);
+    color: var(--vscode-button-foreground);
+    border: none;
+    cursor: pointer;
+    padding: 6px 14px;
+}
+
+button:hover {
+    background-color: var(--vscode-button-hoverBackground);
+}
+
+button:active {
+    opacity: 0.8;
+}
+
+#searchBox {
+    width: 250px;
+}
+
+#layoutSelect {
+    min-width: 150px;
+}
+
+#filterSelect {
+    min-width: 120px;
+}
+
+#graph {
+    flex: 1;
+    border: 1px solid var(--vscode-panel-border);
+    background-color: var(--vscode-editor-background);
+    position: relative;
+    overflow: hidden;
+    min-height: 400px;
+}
+
+#legend {
+    position: absolute;
+    top: 70px;
+    right: 20px;
+    background-color: var(--vscode-sideBar-background);
+    border: 1px solid var(--vscode-panel-border);
+    border-radius: 4px;
+    padding: 12px;
+    font-size: 12px;
+    max-width: 200px;
+    pointer-events: none;
+    z-index: 10;
+}
+
+#legend h3 {
+    margin: 0 0 8px 0;
+    font-size: 13px;
+    font-weight: 600;
+}
+
+.legend-item {
+    display: flex;
+    align-items: center;
+    gap: 8px;
+    margin-bottom: 6px;
+}
+
+.legend-color {
+    width: 16px;
+    height: 16px;
+    border-radius: 50%;
+    border: 1px solid var(--vscode-panel-border);
+}
+
+.legend-label {
+    opacity: 0.9;
+}
+
+#tooltip {
+    position: absolute;
+    background-color: var(--vscode-editorHoverWidget-background);
+    border: 1px solid var(--vscode-editorHoverWidget-border);
+    color: var(--vscode-editorHoverWidget-foreground);
+    padding: 8px 12px;
+    border-radius: 4px;
+    font-size: 12px;
+    pointer-events: none;
+    display: none;
+    z-index: 1000;
+    white-space: pre-line;
+}
+
+#noLinksMessage {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    background-color: var(--vscode-editorWidget-background);
+    border: 2px solid var(--vscode-inputValidation-infoBorder);
+    padding: 24px;
+    border-radius: 6px;
+    max-width: 500px;
+    text-align: center;
+    z-index: 100;
+    pointer-events: auto;
+}
+
+#noLinksMessage h3 {
+    margin: 0 0 12px 0;
+    color: var(--vscode-textLink-foreground);
+}
+
+#noLinksMessage p {
+    margin: 8px 0;
+    line-height: 1.5;
+}
+
+#noLinksMessage code {
+    background-color: var(--vscode-textCodeBlock-background);
+    padding: 2px 6px;
+    border-radius: 3px;
+    font-family: 'Courier New', monospace;
+}
+
+.hidden {
+    display: none !important;
+}
